+++ frontend/domain/models/tasks/span.ts
